refactor: #7 TodoList 컴포넌트 및 기능 분리
